日付の初期入力

+++ src/rakuraku_apps/models.py
@@ -87,17 +87,17 @@ class TankModel(BaseModel):
 
 class WaterQualityModel(BaseModel):
     date = models.DateField("計測日")
-    room_temperature = models.IntegerField("室温", null=True)
-    water_temperature = models.IntegerField("水温", null=True)
-    pH = models.IntegerField("ph", null=True)
-    DO = models.IntegerField("DO", null=True)
-    salinity = models.IntegerField("塩分濃度", null=True)
-    NH4 = models.IntegerField("NH4", null=True)
-    NO2 = models.IntegerField("NO2", null=True)
-    NO3 = models.IntegerField("NO3", null=True)
-    Ca = models.IntegerField("Ca", null=True)
-    Al = models.IntegerField("Al", null=True)
-    Mg = models.IntegerField("Mg", null=True)
+    room_temperature = models.FloatField("室温", null=True)
+    water_temperature = models.FloatField("水温", null=True)
+    pH = models.FloatField("ph", null=True)
+    DO = models.FloatField("DO", null=True)
+    salinity = models.FloatField("塩分濃度", null=True)
+    NH4 = models.FloatField("NH4", null=True)
+    NO2 = models.FloatField("NO2", null=True)
+    NO3 = models.FloatField("NO3", null=True)
+    Ca = models.FloatField("Ca", null=True)
+    Al = models.FloatField("Al", null=True)
+    Mg = models.FloatField("Mg", null=True)
     notes = models.TextField("備考", max_length=512, null=True, blank=True)
     tank = models.ForeignKey(
         "TankModel",
@@ -112,16 +112,16 @@ class WaterQualityModel(BaseModel):
         db_table = "water_quality"
 
 class StandardValueModel(BaseModel):
-    water_temperature = models.IntegerField("水温基準値", null=True)
-    pH = models.IntegerField("pH基準値", null=True)
-    DO = models.IntegerField("DO基準値", null=True)
-    salinity = models.IntegerField("塩分濃度基準値", null=True)
-    NH4 = models.IntegerField("NH4基準値", null=True)
-    NO2 = models.IntegerField("NO2基準値", null=True)
-    NO3 = models.IntegerField("NO3基準値", null=True)
-    Ca = models.IntegerField("Ca基準値", null=True)
-    Al = models.IntegerField("Al基準値", null=True)
-    Mg = models.IntegerField("Mg基準値", null=True)
+    water_temperature = models.FloatField("水温基準値", null=True)
+    pH = models.FloatField("pH基準値", null=True)
+    DO = models.FloatField("DO基準値", null=True)
+    salinity = models.FloatField("塩分濃度基準値", null=True)
+    NH4 = models.FloatField("NH4基準値", null=True)
+    NO2 = models.FloatField("NO2基準値", null=True)
+    NO3 = models.FloatField("NO3基準値", null=True)
+    Ca = models.FloatField("Ca基準値", null=True)
+    Al = models.FloatField("Al基準値", null=True)
+    Mg = models.FloatField("Mg基準値", null=True)
 
     class Meta:
         verbose_name = "基準値"
